--- app/Http/Requests/StorePatient.php
@@ -1,37 +1,0 @@
-<?php
-
-namespace App\Http\Requests;
-
-use App\Patient;
-use Illuminate\Foundation\Http\FormRequest;
-
-class StorePatient extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
-    {
-        return [
-            'firstName' => 'required',
-            'lastName' => 'required'
-        ];
-    }
-
-    public function response(array $errors)
-    {
-        return response()->exception($errors, 400);
-    }
-}
